requirements.txt
requests
python-dotenv
psycopg2-binary
aiogram==3.0.0b7
pandas
pybit==5.7.0
pycryptodome
ccxt


